RE-1191 Use JS implementation of console colours
The ansiColors plugin uses Jenkins' console notes feature to apply
formatting to the console. Unfortunately this adds quite a lot of
data to every log line that uses ansi colour codes. When a Jenkins
job involves executing a large number of playbooks, the number of
console lines that include colour codes is significant.

This pull request provides an alternative, which is a javascript
implementation that converts the ansi codes to html formatting.
As the conversion is done client side, there is no need to store
additional metadata on the server. This means the ansiColor plugin
can be removed, but logs will still be colourised when viewed
in the browser.

Implementation taken from:
 https://github.com/jakubjosef/jenkins-ansi-color-js